Added file list shortcuts
