+++ packages/code-studio/src/settings/Shortcuts.ts
@@ -25,25 +25,8 @@ const NOTEBOOK = {
   }),
 };
 
-const FILE_EXPLORER = {
-  name: 'File Explorer',
-  DELETE: ShortcutRegistry.createAndAdd({
-    id: 'FILE_EXPLORER.DELETE',
-    name: 'Delete',
-    shortcut: [KEY.DELETE],
-    macShortcut: [MODIFIER.CMD, KEY.BACKSPACE],
-  }),
-  RENAME: ShortcutRegistry.createAndAdd({
-    id: 'FILE_EXPLORER.RENAME',
-    name: 'Rename',
-    shortcut: [KEY.ENTER],
-    macShortcut: [KEY.ENTER],
-  }),
-};
-
 const SHORTCUTS = {
   NOTEBOOK,
-  FILE_EXPLORER,
 };
 
 export default SHORTCUTS;
